[PIR #859] Extract Markdown Preview webview HTML into preview-template.ts
